+++ R/data.R
@@ -1,0 +1,7 @@
+#' Sample data of clipboard bitmap array
+#'
+#' @format raw data of bitmap array with 246292 length
+#' @examples
+#' data(clipboard_sample)
+#' 
+"clipboard_sample"