+++ R/utils-rlang.R
@@ -1,5 +1,2 @@
 #' @importFrom rlang :=
 NULL
-
-#' @importFrom rlang .data
-NULL