+++ NEWS.md
@@ -1,7 +1,13 @@
+# screenshot 0.9.1
+
+* 2024-05-30
+    * `display_corner()` to locate image in the display corner.
+    * `save_clipboard_image()` to save clipboard image as a png file.
+
 # screenshot 0.9.0
 
 * 2023-08-11
 
 * First release
-    * `screenshot()` to take a screenshot
+    * `screenshot()` to take a screenshot.
     * `locate_image()` to locate a image position on a screenshot image.